Stop tracking application.properties and create staysavedto

+++ plax-backend/src/main/java/com/maxdev/plaxbackend/modules/Stay/Service/Impl/StayService.java
@@ -3,35 +3,41 @@ package com.maxdev.plaxbackend.modules.Stay.Service.Impl;
 import com.maxdev.plaxbackend.modules.Exception.ResourceAlreadyExistsException;
 import com.maxdev.plaxbackend.modules.Exception.ResourceNotFoundException;
 import com.maxdev.plaxbackend.modules.Stay.DTO.StayDTO;
+import com.maxdev.plaxbackend.modules.Stay.DTO.StaySaveDTO;
 import com.maxdev.plaxbackend.modules.Stay.Mapper.StayMapper;
+import com.maxdev.plaxbackend.modules.Stay.Mapper.StaySaveMapper;
 import com.maxdev.plaxbackend.modules.Stay.Repository.StayImageRepository;
 import com.maxdev.plaxbackend.modules.Stay.Repository.StayRepository;
 import com.maxdev.plaxbackend.modules.Stay.Service.IStayService;
 import com.maxdev.plaxbackend.modules.Stay.Stay;
 import com.maxdev.plaxbackend.modules.Stay.StayImage;
 
+import com.maxdev.plaxbackend.modules.Util.BaseUrl;
 import lombok.extern.log4j.Log4j2;
 
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.UrlResource;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.io.IOException;
+import java.net.MalformedURLException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
+import java.util.stream.Collectors;
 
 import static java.time.LocalDateTime.now;
 
 @Log4j2
 @Service
-public class StayService implements IStayService {
+public class StayService implements IStayService, BaseUrl {
 
     /*
      * private final: La variable debe ser inicializada una vez y no puede ser
@@ -48,51 +54,113 @@ public class StayService implements IStayService {
 
     @Override
     @Transactional
-    public StayDTO save(StayDTO stayDTO) {
+    public StaySaveDTO save(StaySaveDTO stayDTO, MultipartFile[] images) throws IOException {
         stayRepository.findByName(stayDTO.getName()).ifPresent(stay -> {
             log.error("Stay with name: {} already exists", stayDTO.getName());
             throw new ResourceAlreadyExistsException("Stay with name: " + stayDTO.getName() + " already exists");
         });
-        Stay stayToSave = StayMapper.INSTANCE.dtoToEntity(stayDTO);
+        Set<String> imageNames = saveImages(images);
+        stayDTO.setImages(imageNames);
+        Stay stayToSave = StaySaveMapper.INSTANCE.dtoToEntity(stayDTO);
         stayRepository.save(stayToSave);
         log.info("Stay saved: {}", stayToSave.getName());
-        return StayMapper.INSTANCE.entityToDto(stayToSave);
+        return StaySaveMapper.INSTANCE.entityToDto(stayToSave);
     }
 
     @Override
     @Transactional
-    public StayDTO update(StayDTO stayDTO) throws ResourceNotFoundException {
-        stayRepository.findById(stayDTO.getId()).orElseThrow(() -> new ResourceNotFoundException("Stay not found with id: " + stayDTO.getId()));
+    public StaySaveDTO update(StaySaveDTO stayDTO, MultipartFile[] images, Set<String> imagesToDelete) throws IOException {
+        log.debug("Updating stay: {}", stayDTO.getId());
+        boolean hasImageToDelete = imagesToDelete != null && !imagesToDelete.isEmpty();
+        Stay stayToUpdate = stayRepository.findById(stayDTO.getId()).orElseThrow(() -> {
+            log.error("Stay with id: {} not found", stayDTO.getId());
+            return new ResourceNotFoundException("Stay with id: " + stayDTO.getId() + " not found");
+        });
 
-        Stay stayToUpdate = StayMapper.INSTANCE.dtoToEntity(stayDTO);
+        if (!stayToUpdate.getName().equals(stayDTO.getName())) {
+            stayRepository.findByName(stayDTO.getName()).ifPresent(stay -> {
+                log.error("Stay with name: {} already exists", stayDTO.getName());
+                throw new ResourceAlreadyExistsException("Stay with name: " + stayDTO.getName() + " already exists");
+            });
+        }
 
+        stayDTO.setImages(stayToUpdate.getImages().stream().map(StayImage::getUrl).collect(Collectors.toSet()));
+        if (images != null) {
+            Set<String> imageNames = saveImages(images);
+            stayDTO.getImages().addAll(imageNames);
+        }
+
+        if (hasImageToDelete) {
+            stayDTO.getImages().removeAll(imagesToDelete);
+        }
+
+        // Reutiliza las imagenes que ya existen en la base de datos
+        stayToUpdate = StaySaveMapper.INSTANCE.dtoToEntity(stayDTO);
         for (StayImage stayImage : stayToUpdate.getImages()) {
             if (stayImageRepository.findByUrl(stayImage.getUrl()).isPresent())
                 stayImage.setId(stayImageRepository.findByUrl(stayImage.getUrl()).get().getId());
         }
 
-        stayRepository.save(stayToUpdate);
+        stayToUpdate = stayRepository.save(stayToUpdate);
+
+        if (hasImageToDelete) {
+            stayImageRepository.deleteByUrls(imagesToDelete);
+        }
+
         log.info("Stay updated: {}", stayToUpdate.getName());
-        return StayMapper.INSTANCE.entityToDto(stayToUpdate);
+        return StaySaveMapper.INSTANCE.entityToDto(stayToUpdate);
     }
 
     @Override
     @Transactional
-    public Optional<StayDTO> findById(UUID id) throws ResourceNotFoundException {
-        return stayRepository.findById(id).map(StayMapper.INSTANCE::entityToDto);
+    public StayDTO findById(UUID id) throws ResourceNotFoundException {
+        Stay stay = stayRepository.findById(id)
+                .orElseThrow(() -> {
+                    log.error("Stay with id: {} not found", id);
+                    return new ResourceNotFoundException("Stay with id: " + id + " not found");
+                });
+        StayDTO stayFound = StayMapper.INSTANCE.entityToDto(stay);
+        stayFound.setImages(stayFound.getImages().stream()
+                .map(image -> getBaseUrl() + "/api/stays/images/" + image)
+                .collect(Collectors.toSet()));
+        stayFound.setFeatures(stayFound.getFeatures().stream()
+                .peek(feature -> feature.setIcon(getBaseUrl() + "/api/features/svg/" + feature.getIcon()))
+                .collect(Collectors.toSet()));
+        return stayFound;
     }
 
     @Override
     @Transactional
-    public Optional<StayDTO> findByName(String name) {
-        return stayRepository.findByName(name).map(StayMapper.INSTANCE::entityToDto);
+    public StayDTO findByName(String name) {
+        log.debug("Finding stay by name: {}", name);
+        Stay stay = stayRepository.findByName(name).orElseThrow(() -> {
+            log.error("Stay with name: {} not found", name);
+            return new ResourceNotFoundException("Stay with name: " + name + " not found");
+        });
+        StayDTO stayFound = StayMapper.INSTANCE.entityToDto(stay);
+        stayFound.setImages(stayFound.getImages().stream()
+                .map(image -> getBaseUrl() + "/api/stays/images/" + image)
+                .collect(Collectors.toSet()));
+        stayFound.setFeatures(stayFound.getFeatures().stream()
+                .peek(feature -> feature.setIcon(getBaseUrl() + "/api/features/svg/" + feature.getIcon()))
+                .collect(Collectors.toSet()));
+        return stayFound;
     }
 
     @Override
     @Transactional
     public Page<StayDTO> findAll(Pageable pageable) {
         log.debug("Finding all stays with pageable: {}", pageable);
-        return stayRepository.findAll(pageable).map(StayMapper.INSTANCE::entityToDto);
+        Page<StayDTO> pageStays = stayRepository.findAll(pageable).map(StayMapper.INSTANCE::entityToDto);
+        pageStays.forEach(stayDTO -> {
+            stayDTO.setImages(stayDTO.getImages().stream()
+                    .map(image -> getBaseUrl() + "/api/stays/images/" + image)
+                    .collect(Collectors.toSet()));
+            stayDTO.setFeatures(stayDTO.getFeatures().stream()
+                    .peek(feature -> feature.setIcon(getBaseUrl() + "/api/features/svg/" + feature.getIcon()))
+                    .collect(Collectors.toSet()));
+        });
+        return pageStays;
     }
 
     @Override
@@ -101,24 +169,54 @@ public class StayService implements IStayService {
         log.debug("Finding random stays with size: {}", size);
         Set<StayDTO> randomStays = new HashSet<>();
         List<Stay> stays = stayRepository.findAll();
-        if (stays.size() < size) size = stays.size();
+        if (stays.size() < size)
+            size = stays.size();
         Random random = new Random();
         while (randomStays.size() < size)
             randomStays.add(StayMapper.INSTANCE.entityToDto(stays.get(random.nextInt(stays.size()))));
+
+        randomStays.forEach(stay -> {
+            stay.setImages(stay.getImages().stream()
+                    .map(image -> getBaseUrl() + "/api/stays/images/" + image)
+                    .collect(Collectors.toSet()));
+            stay.setFeatures(stay.getFeatures().stream()
+                    .peek(feature -> feature.setIcon(getBaseUrl() + "/api/features/svg/" + feature.getIcon()))
+                    .collect(Collectors.toSet()));
+        });
+        log.info("Returning {} random stays", randomStays.size());
         return randomStays;
     }
 
     @Override
     @Transactional
     public StayDTO delete(UUID id) throws ResourceNotFoundException, IOException {
-        Stay stayToDelete = stayRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Stay not found with id: " + id));
+        log.debug("Deleting stay with id: {}", id);
+        Stay stayToDelete = stayRepository.findById(id)
+                .orElseThrow(() -> {
+                    log.error("Stay with id: {} not found", id);
+                    return new ResourceNotFoundException("Stay with id: " + id + " not found");
+                });
 
         stayRepository.delete(stayToDelete);
-        deleteFileImages(stayToDelete.getImages());
+        if (!stayRepository.existsById(id)) {
+            deleteFileImages(stayToDelete.getImages());
+        }
+        log.info("Stay deleted: {}", stayToDelete.getName());
         return StayMapper.INSTANCE.entityToDto(stayToDelete);
     }
 
-    public Set<String> saveImages(MultipartFile[] images) throws IOException {
+    public Resource getImage(String imageName) throws MalformedURLException {
+        log.debug("Getting image with name: {}", imageName);
+        Path imagePath = Paths.get("uploads/stays").resolve(imageName);
+        Resource resource = new UrlResource(imagePath.toUri());
+        if (!resource.exists()) {
+            log.error("Image not found: {}", imageName);
+            throw new ResourceNotFoundException("Image not found: " + imageName);
+        }
+        return resource;
+    }
+
+    private Set<String> saveImages(MultipartFile[] images) throws IOException {
         Set<String> imageNames = new HashSet<>();
         String uploadDir = "uploads/stays";
         Path uploadPath = Paths.get(uploadDir);
@@ -135,19 +233,10 @@ public class StayService implements IStayService {
         return imageNames;
     }
 
-    public void deleteFileImages(Set<StayImage> imageNames) throws IOException {
+    private void deleteFileImages(Set<StayImage> imageNames) throws IOException {
         for (StayImage imageName : imageNames) {
             Path filePath = Paths.get("uploads/stays").resolve(imageName.getUrl()).normalize();
             Files.deleteIfExists(filePath);
         }
-    }
-
-    public void deleteImages(Set<String> imageNames) throws IOException {
-
-        stayImageRepository.deleteByUrls(imageNames);
-    }
-
-    public String getBaseUrl() {
-        return ServletUriComponentsBuilder.fromCurrentContextPath().path("/api/stays/").toUriString();
     }
 }
